# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skillbytes API - PBL Module",
        "description": "REST API for Problem-Based Learning (PBL) Case Management System",
        "contact": {
            "email": "support@skillbytes.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development"
        },
        {
            "url": "http://145.79.13.180/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/mata-pelajarans": {
            "get": {
                "tags": [
                    "MataPelajaran"
                ],
                "summary": "Get paginated mata pelajarans",
                "description": "Retrieve paginated list of learning subjects (mata pelajaran).",
                "operationId": "efb07975aa600fa5cf2da4183a920eb1",
                "parameters": [
                    {
                        "name": "pagesize",
                        "in": "query",
                        "description": "Page size (default: 15)",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated mata pelajarans",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mata pelajarans retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "MataPelajaran"
                ],
                "summary": "Create new mata pelajaran",
                "description": "Create a new learning subject. Only accessible to admin and guru roles.",
                "operationId": "58ed41ffd034e6bf0740df4691d7a5a0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "level_number"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Mata pelajaran name",
                                        "type": "string"
                                    },
                                    "level_number": {
                                        "description": "Mata pelajaran number",
                                        "type": "integer"
                                    },
                                    "description": {
                                        "description": "Mata pelajaran description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mata pelajaran created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mata pelajaran created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mata-pelajarans/all": {
            "get": {
                "tags": [
                    "MataPelajaran"
                ],
                "summary": "Get all mata pelajarans with lessons",
                "description": "Retrieve all learning subjects with their associated lessons.",
                "operationId": "a9e55ab6c27d7d5e74fc3c99a95868fa",
                "responses": {
                    "200": {
                        "description": "All mata pelajarans with lessons",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All mata pelajarans retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mata-pelajarans/{mataPelajaran}": {
            "get": {
                "tags": [
                    "MataPelajaran"
                ],
                "summary": "Get mata pelajaran by number",
                "description": "Retrieve detailed information of a specific mata pelajaran by level number.",
                "operationId": "87dc20637c4a8c1e95eebd3c5b1635d9",
                "parameters": [
                    {
                        "name": "mataPelajaran",
                        "in": "path",
                        "description": "Mata pelajaran number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mata pelajaran retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mata pelajaran retrieved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Mata pelajaran not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "MataPelajaran"
                ],
                "summary": "Update mata pelajaran",
                "description": "Update an existing learning subject. Only accessible to admin and guru roles.",
                "operationId": "9806ec01650b233a5724a0c5154d08e6",
                "parameters": [
                    {
                        "name": "mataPelajaran",
                        "in": "path",
                        "description": "Mata pelajaran number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Updated description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mata pelajaran updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mata pelajaran updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Mata pelajaran not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "MataPelajaran"
                ],
                "summary": "Delete mata pelajaran",
                "description": "Delete a learning subject permanently. Only accessible to admin and guru roles.",
                "operationId": "a6b4faac8f5eaaeecac0c5e6bbf460f1",
                "parameters": [
                    {
                        "name": "mataPelajaran",
                        "in": "path",
                        "description": "Mata pelajaran number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mata pelajaran deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mata pelajaran deleted successfully"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Mata pelajaran not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token for API authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}